Correct target file suffixes.
Correct target file suffixes for "chunked", "man", "htmlhelp" and "text" based
on the a2x man page.  Due to the case of "chunked", the target_factory needed to
be set to "Entry", since it produces a directory as output.  Also, the change to
'text' avoids a collision with the source file name.