Add A2X_KEEPARTIFACTS var that maps to -k option.
Add a A2X_KEEPARTIFACTS construction variable that maps to the a2x -k option.
The advantage: the build can be set to not be re-triggered when this option
changes.  It also makes it clear that this option is somewhat special: you
should typically leave it at its default.  It is only there "just in case".