Reorganise code into submodules.
The code for this tool has gotten somewhat difficult to follow for a mere ~550
lines.  I think it makes sense to move most of it into submodules:

- builders.py
- pseudo_builders.py

They define the actual builders and the pseudo-builders, respectively.  The
__init__.py file now only defines generate(), exists() and some helper code.

In addition to simply moving code to the new submodules, it has been
restructured somewhat to make it easier to follow.