Catch all exceptions when asciidoc/a2x not found.
Catch all exceptions when the asciidoc/a2x executable cannot be found by using a
simple "except:" instead of "except (OSError, AttributeError):" to increase
robustness.  The motivation is that on Windows, you get a TypeError instead of
an AttributeError (like on Linux) when passing None as an executable name to
subprocess.Popen().  So, to take care of other unexpected situations, just use
"except:", instead of "except (list, of, exceptions):".