Add file/dir names in A2X_FLAGS to dependencies.
The A2X builder now looks for file/dir names in A2X_FLAGS and adds them to the
dependency list if they exist.  This handles more exotic options in a generic
way.  For example, the reason I added this in the first place was that I wanted
latex style files passed via '--dblatex-opts "-s <style_file>"' to be added to
the dependency list without having to add a special construction variable for
it, or for other, similar options for that matter.