Fix the macro regexps to not match whitespace.
Fix the image and include macro regexps to not match whitespace in the macro
target, since whitespace is not allowed according to the AsciiDoc documentation.
This caused problems when

- the opening square bracket is left out (which is a syntax error in AsciiDoc),
- and the line contained an opening square bracket afterwards.

Under these circumstances the regexp would match everything following the actual
target up to the opening bracket, which was then considered a file name.
Naturally this caused "file not found" type errors.

This commit fixes issue #1.